Add support for PBKDF2 for blind indexing
For very sensitive information (Social Security numbers, tax
identification numbers, or other ostensibly unique numbers assigned to
people), it is a good idea to use a key stretching algorithm for the
blind index to make it harder to brute force the hashed values. PBKDF2
is a key stretching algorithm that accomplishes this.