Migration task succeeds on empty table
This functionality appears to have been introduced in
cdeae97c69f7c54aeadf0921db30dc411d7a9938; however, it looks like it was
removed as part of the refactor to use a cursor for IDs rather than
requiring numeric IDs in 01946dcd39fe96be9b79c1d095fc71ec311cfd93.

With this change, if the migrator is passed a schema that is linked to
an empty table, an error is no longer raised. Instead, the migration is
not attempted, and the function returns that the migration completed.

This does have the impact of changing the return value of the function
in this scenario. Rather than the `:ok` it otherwise returns, instead
it returns a tuple `{:ok, message}`, where the message includes
information about why no migration took place. It doesn't appear that
the return value of this function is used; however, this technically
breaks the public API in the scenario where the schema has no data.

While that tuple is not surfaced to the end-user as part of this effort,
it has the impact of making testing this change easier by making the
impact observable for something that would have no side-effect
(given that the migration does nothing, as there's no data to migrate).
Presumably this could be helpful to callers as well.